--- a/subprojects/core-api/src/main/java/org/gradle/api/attributes/AttributeCompatibilityRule.java
+++ b/subprojects/core-api/src/main/java/org/gradle/api/attributes/AttributeCompatibilityRule.java
@@ -26,7 +26,7 @@ import org.gradle.api.Incubating;
  * mark the producer value as compatible or not.
  * <p>
  * Note that the rule will never receive a {@code CompatibilityCheckDetails} that has {@code equal} consumer and producer
- * values as this check is performed before invoking the rule and assumes compatiblity in that case.
+ * values as this check is performed before invoking the rule and assumes compatibility in that case.
  *
  * @since 4.0
  * @param <T> The attribute value type.