--- a/subprojects/core-api/src/main/java/org/gradle/api/attributes/Category.java
+++ b/subprojects/core-api/src/main/java/org/gradle/api/attributes/Category.java
@@ -23,10 +23,11 @@ import org.gradle.internal.scan.UsedByScanPlugin;
 /**
  * This attribute describes the categories of variants for a given module.
  * <p>
- * Two values are found in published components:
+ * Three values are found in published components:
  * <ul>
  *     <li>{@code library}: Indicates that the variant is a library, that usually means a binary and a set of dependencies</li>
- *     <li>{@code platform}: indicates that the variant is a platform, that usually means a definition of dependency constraints</li>
+ *     <li>{@code platform}: Indicates that the variant is a platform, that usually means a definition of dependency constraints</li>
+ *     <li>{@code documentation}: Indicates that the variant is documentation of the software module</li>
  * </ul>
  * One value is used for derivation. A {@code platform} variant can be consumed as a {@code enforced-platform} which means all the dependency
  * information it provides is applied as {@code forced}.
@@ -53,4 +54,12 @@ public interface Category extends Named {
      * The enforced platform, usually a synthetic variant derived from the {@code platform}
      */
     String ENFORCED_PLATFORM = "enforced-platform";
+
+    /**
+     * The documentation category
+     *
+     * @since 5.6
+     */
+    String DOCUMENTATION = "documentation";
+
 }