--- a/subprojects/core/src/main/java/org/gradle/api/internal/artifacts/JavaEcosystemSupport.java
+++ b/subprojects/core/src/main/java/org/gradle/api/internal/artifacts/JavaEcosystemSupport.java
@@ -75,7 +75,9 @@ public abstract class JavaEcosystemSupport {
             @Override
             public void execute(ActionConfiguration actionConfiguration) {
                 actionConfiguration.params(objectFactory.named(Usage.class, Usage.JAVA_API));
+                actionConfiguration.params(objectFactory.named(Usage.class, Usage.JAVA_API_JARS));
                 actionConfiguration.params(objectFactory.named(Usage.class, Usage.JAVA_RUNTIME));
+                actionConfiguration.params(objectFactory.named(Usage.class, Usage.JAVA_RUNTIME_JARS));
             }
         });
     }
@@ -92,13 +94,23 @@ public abstract class JavaEcosystemSupport {
     public static class UsageDisambiguationRules implements AttributeDisambiguationRule<Usage>, ReusableAction {
         final Usage javaApi;
         final Usage javaRuntime;
+        final Usage javaApiJars;
+        final Usage javaRuntimeJars;
 
+        final ImmutableSet<Usage> apiVariants;
+        final ImmutableSet<Usage> runtimeVariants;
 
         @Inject
         UsageDisambiguationRules(Usage javaApi,
-                                 Usage javaRuntime) {
+                                 Usage javaApiJars,
+                                 Usage javaRuntime,
+                                 Usage javaRuntimeJars) {
             this.javaApi = javaApi;
+            this.javaApiJars = javaApiJars;
+            this.apiVariants = ImmutableSet.of(javaApi, javaApiJars);
             this.javaRuntime = javaRuntime;
+            this.javaRuntimeJars = javaRuntimeJars;
+            this.runtimeVariants = ImmutableSet.of(javaRuntime, javaRuntimeJars);
         }
 
         @Override
@@ -106,17 +118,30 @@ public abstract class JavaEcosystemSupport {
             Set<Usage> candidateValues = details.getCandidateValues();
             Usage consumerValue = details.getConsumerValue();
             if (consumerValue == null) {
-                if (candidateValues.contains(javaRuntime)) {
+                if (candidateValues.contains(javaRuntimeJars)) {
+                    // Use the Jars when nothing has been requested
+                    details.closestMatch(javaRuntimeJars);
+                } else if (candidateValues.contains(javaRuntime)) {
                     // Use the runtime when nothing has been requested
                     details.closestMatch(javaRuntime);
                 }
             } else {
-                if (candidateValues.contains(consumerValue)) {
+                if (javaRuntime.equals(consumerValue)) {
+                    // we're asking for a runtime variant, prefer -jars first
+                    if (candidateValues.contains(javaRuntimeJars)) {
+                        details.closestMatch(javaRuntimeJars);
+                    } else if (candidateValues.contains(javaRuntime)) {
+                        details.closestMatch(javaRuntime);
+                    }
+                } else if (candidateValues.contains(consumerValue)) {
                     details.closestMatch(consumerValue);
                 } else if (javaApi.equals(consumerValue)) {
-                    // we're asking for an API variant, but no exact match was found
-                    if (candidateValues.contains(javaRuntime)) {
-                        // java-runtime is a match
+                    // we're asking for an API variant, prefer -jars first for runtime
+                    if (candidateValues.contains(javaApiJars)) {
+                        details.closestMatch(javaApiJars);
+                    } else if (candidateValues.contains(javaRuntimeJars)) {
+                        details.closestMatch(javaRuntimeJars);
+                    } else if (candidateValues.contains(javaRuntime)) {
                         details.closestMatch(javaRuntime);
                     }
                 }
@@ -126,14 +151,24 @@ public abstract class JavaEcosystemSupport {
 
     @VisibleForTesting
     public static class UsageCompatibilityRules implements AttributeCompatibilityRule<Usage>, ReusableAction {
+        private static final Set<String> COMPATIBLE_WITH_JAVA_API = ImmutableSet.of(
+                Usage.JAVA_API_JARS,
+                Usage.JAVA_RUNTIME_JARS,
+                Usage.JAVA_RUNTIME
+        );
         @Override
         public void execute(CompatibilityCheckDetails<Usage> details) {
             String consumerValue = details.getConsumerValue().getName();
             String producerValue = details.getProducerValue().getName();
             if (consumerValue.equals(Usage.JAVA_API)) {
-                if (Usage.JAVA_RUNTIME.equals(producerValue)) {
+                if (COMPATIBLE_WITH_JAVA_API.contains(producerValue)) {
                     details.compatible();
                 }
+                return;
+            }
+            if (consumerValue.equals(Usage.JAVA_RUNTIME) && producerValue.equals(Usage.JAVA_RUNTIME_JARS)) {
+                details.compatible();
+                return;
             }
         }
     }