--- a/subprojects/core/src/main/java/org/gradle/api/internal/artifacts/JavaEcosystemSupport.java
+++ b/subprojects/core/src/main/java/org/gradle/api/internal/artifacts/JavaEcosystemSupport.java
@@ -75,12 +75,7 @@ public abstract class JavaEcosystemSupport {
             @Override
             public void execute(ActionConfiguration actionConfiguration) {
                 actionConfiguration.params(objectFactory.named(Usage.class, Usage.JAVA_API));
-                actionConfiguration.params(objectFactory.named(Usage.class, Usage.JAVA_API_JARS));
-                actionConfiguration.params(objectFactory.named(Usage.class, Usage.JAVA_API_CLASSES));
                 actionConfiguration.params(objectFactory.named(Usage.class, Usage.JAVA_RUNTIME));
-                actionConfiguration.params(objectFactory.named(Usage.class, Usage.JAVA_RUNTIME_JARS));
-                actionConfiguration.params(objectFactory.named(Usage.class, Usage.JAVA_RUNTIME_CLASSES));
-                actionConfiguration.params(objectFactory.named(Usage.class, Usage.JAVA_RUNTIME_RESOURCES));
             }
         });
     }
@@ -97,32 +92,13 @@ public abstract class JavaEcosystemSupport {
     public static class UsageDisambiguationRules implements AttributeDisambiguationRule<Usage>, ReusableAction {
         final Usage javaApi;
         final Usage javaRuntime;
-        final Usage javaApiJars;
-        final Usage javaApiClasses;
-        final Usage javaRuntimeJars;
-        final Usage javaRuntimeClasses;
-        final Usage javaRuntimeResources;
 
-        final ImmutableSet<Usage> apiVariants;
-        final ImmutableSet<Usage> runtimeVariants;
 
         @Inject
         UsageDisambiguationRules(Usage javaApi,
-                                 Usage javaApiJars,
-                                 Usage javaApiClasses,
-                                 Usage javaRuntime,
-                                 Usage javaRuntimeJars,
-                                 Usage javaRuntimeClasses,
-                                 Usage javaRuntimeResources) {
+                                 Usage javaRuntime) {
             this.javaApi = javaApi;
-            this.javaApiJars = javaApiJars;
-            this.javaApiClasses = javaApiClasses;
-            this.apiVariants = ImmutableSet.of(javaApi, javaApiJars, javaApiClasses);
             this.javaRuntime = javaRuntime;
-            this.javaRuntimeJars = javaRuntimeJars;
-            this.javaRuntimeClasses = javaRuntimeClasses;
-            this.javaRuntimeResources = javaRuntimeResources;
-            this.runtimeVariants = ImmutableSet.of(javaRuntime, javaRuntimeJars, javaRuntimeClasses, javaRuntimeResources);
         }
 
         @Override
@@ -130,10 +106,7 @@ public abstract class JavaEcosystemSupport {
             Set<Usage> candidateValues = details.getCandidateValues();
             Usage consumerValue = details.getConsumerValue();
             if (consumerValue == null) {
-                if (candidateValues.contains(javaRuntimeJars)) {
-                    // Use the Jars when nothing has been requested
-                    details.closestMatch(javaRuntimeJars);
-                } else if (candidateValues.contains(javaRuntime)) {
+                if (candidateValues.contains(javaRuntime)) {
                     // Use the runtime when nothing has been requested
                     details.closestMatch(javaRuntime);
                 }
@@ -142,24 +115,8 @@ public abstract class JavaEcosystemSupport {
                     details.closestMatch(consumerValue);
                 } else if (javaApi.equals(consumerValue)) {
                     // we're asking for an API variant, but no exact match was found
-                    if (candidateValues.contains(javaApiClasses)) {
-                        // prefer the most lightweight API
-                        details.closestMatch(javaApiClasses);
-                    } else if (candidateValues.contains(javaApiJars)) {
-                        details.closestMatch(javaApiJars);
-                    } else if (candidateValues.contains(javaApi)) {
-                        // Prefer the API over the runtime when the API has been requested
-                        details.closestMatch(javaApi);
-                    } else if (candidateValues.contains(javaRuntimeClasses)) {
-                        details.closestMatch(javaRuntimeClasses);
-                    } else if (candidateValues.contains(javaRuntimeJars)) {
-                        details.closestMatch(javaRuntimeJars);
-                    }
-                } else if (javaRuntime.equals(consumerValue)) {
-                    // we're asking for a runtime variant, but no exact match was found
-                    if (candidateValues.contains(javaRuntimeJars)) {
-                        details.closestMatch(javaRuntimeJars);
-                    } else if (candidateValues.contains(javaRuntime)) {
+                    if (candidateValues.contains(javaRuntime)) {
+                        // java-runtime is a match
                         details.closestMatch(javaRuntime);
                     }
                 }
@@ -169,26 +126,14 @@ public abstract class JavaEcosystemSupport {
 
     @VisibleForTesting
     public static class UsageCompatibilityRules implements AttributeCompatibilityRule<Usage>, ReusableAction {
-        private static final Set<String> COMPATIBLE_WITH_JAVA_API = ImmutableSet.of(
-                Usage.JAVA_API_JARS,
-                Usage.JAVA_API_CLASSES,
-                Usage.JAVA_RUNTIME_JARS,
-                Usage.JAVA_RUNTIME_CLASSES,
-                Usage.JAVA_RUNTIME
-        );
         @Override
         public void execute(CompatibilityCheckDetails<Usage> details) {
             String consumerValue = details.getConsumerValue().getName();
             String producerValue = details.getProducerValue().getName();
             if (consumerValue.equals(Usage.JAVA_API)) {
-                if (COMPATIBLE_WITH_JAVA_API.contains(producerValue)) {
+                if (Usage.JAVA_RUNTIME.equals(producerValue)) {
                     details.compatible();
                 }
-                return;
-            }
-            if (consumerValue.equals(Usage.JAVA_RUNTIME) && producerValue.equals(Usage.JAVA_RUNTIME_JARS)) {
-                details.compatible();
-                return;
             }
         }
     }