A solution for #6784.

This PR requires #9670 to be merged first.

### How it works

Turn on using the `jar`, built by the _jar task_, instead of the class folder on the compile classpath by setting the following system property:

`org.gradle.java.compile-classpath-packaging=true`

### Measurements 
Results for measuring the [huge project](https://github.com/gradle/gradle/blob/5b7a84089e44e9eee446ee7401ff7f6fb4a44fda/subprojects/internal-performance-testing/src/main/groovy/org/gradle/performance/generator/JavaTestProject.groovy#L38-L41) on Windows:

**Without compile-classpath-packaging** 

```
    Measuring Gradle tasks: clean classes
    Speed java-library-plugin: Results were inconclusive with 95% confidence.
    Difference: 22.622 s slower (22622 ms), 9.79%
      java-library-plugin median: 4.23 m min: 4.164 m, max: 4.266 m, se: 2.516 s}
      > [4.266 m, 4.23 m, 4.164 m]
      java-plugin median: 3.853 m min: 3.837 m, max: 3.866 m, se: 710.32 ms}
      > [3.837 m, 3.866 m, 3.853 m]
```

**With compile-classpath-packaging** 
```
    Measuring Gradle tasks: clean classes
    Speed java-library-plugin: Results were inconclusive with 95% confidence.
    Difference: 10.114 s faster (10114 ms), -4.37%
      java-library-plugin median: 3.688 m min: 3.671 m, max: 3.72 m, se: 1.22 s}
      > [3.688 m, 3.671 m, 3.72 m]
      java-plugin median: 3.856 m min: 3.839 m, max: 3.86 m, se: 544.53 ms}
      > [3.839 m, 3.86 m, 3.856 m]
```