@big-guy 2019-06-18
Isn't this just recreating the regular jar without resources?  Does the lack of resources explain why it's 10s faster than just the Java plugin?
If there are performance improvements that could be made to the Java plugin, the advice could just be to use the Java plugin when a project is over a certain size?