@big-guy 2019-06-18
Thanks @jjohannes.  I was just curious.

Would it make sense to just make this the way it works instead of using a property?  e.g., on Linux/mac, we always want classes-dir packaging and on Windows, we always want jar packaging because we know it's faster and just let variant dependency management pick the right variant?

One interesting wrinkle with this is how it affects the build cache.  Would two builds (one on Linux and one on Windows) with otherwise identical inputs (only difference is classes vs jars) be able to pull the other's result from the cache? If the answer is "no", then that's a step back from what they can do today, I think.